backend/requirements.txt
asgiref==3.2.10
attrs==20.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
factory-boy==3.0.1
Faker==4.1.2
flake8==3.8.3
idna==2.10
inflection==0.5.0
iniconfig==1.0.1
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pip-compile-multi==2.1.0
pip-tools==5.3.1
pluggy==0.13.1
prompt-toolkit==3.0.6
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytest-factoryboy==2.0.3
pytest-mock==3.2.0
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
toposort==1.5
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
